pyproject.toml
[project]
name = "mcp-test-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "google-adk>=0.0.1",
    "ruff>=0.11.6",
]
